reenable lost tests
