+++ system/t01_version/__init__.py
@@ -1,14 +1,3 @@
 """
 Test aptly version
 """
-
-from lib import BaseTest
-
-
-class VersionTest(BaseTest):
-    """
-    version should match
-    """
-    gold_processor = BaseTest.expand_environ
-
-    runCmd = "aptly version"
