+++ system/t01_version/version.py
@@ -1,0 +1,14 @@
+"""
+Test aptly version
+"""
+
+from lib import BaseTest
+
+
+class VersionTest(BaseTest):
+    """
+    version should match
+    """
+    gold_processor = BaseTest.expand_environ
+
+    runCmd = "aptly version"
